+++ src/renderer/stores/chat.ts
@@ -128,6 +128,18 @@ chatStore.attachReducer(({ type, payload, id }, state) => {
       },
     }
     return { ...state, messages }
+  } else if (type === 'MESSAGE_FAILED') {
+    const messages = {
+      ...state.messages,
+      [payload]: {
+        ...state.messages[payload],
+        msg: {
+          ...state.messages[payload].msg,
+          status: 'error',
+        },
+      },
+    }
+    return { ...state, messages }
   } else if (type === 'MESSAGE_READ') {
     const messages = {
       ...state.messages,
@@ -241,6 +253,9 @@ chatStore.attachEffect(async ({ type, payload }, state) => {
       payload[0],
       payload[1]
     )
+
+    // Workaround for failed messages
+    if (messageObj[0] === 0) return
     chatStore.dispatch({
       type: 'MESSAGE_SENT',
       payload: messageObj,
@@ -280,6 +295,28 @@ ipcBackend.on('DC_EVENT_MSG_DELIVERED', (_evt, [id, msgId]) => {
   chatStore.dispatch({
     type: 'MESSAGE_DELIVERED',
     id,
+    payload: msgId,
+  })
+})
+
+ipcBackend.on('DC_EVENT_MSG_FAILED', async (_evt, [chatId, msgId]) => {
+  const state = chatStore.getState()
+  if (state.id !== chatId) return
+  if (!state.messageIds.includes(msgId)) {
+    // Hacking around https://github.com/deltachat/deltachat-desktop/issues/1361#issuecomment-776291299
+    const messageObj = [
+      msgId,
+      await DeltaBackend.call('messageList.getMessage', msgId),
+    ]
+    chatStore.dispatch({
+      type: 'MESSAGE_SENT',
+      payload: messageObj,
+      id: chatId,
+    })
+  }
+  chatStore.dispatch({
+    type: 'MESSAGE_FAILED',
+    id: chatId,
     payload: msgId,
   })
 })