+++ CHANGELOG.md
@@ -44,6 +44,7 @@
 - Fix copy of labeled links
 - Fix hard to alter text in the middle of "Video chat instance" input field (see #2016)
 - Add missing options (at once, one minute...) for autodelete from server
+- Update failed messages also in messagelist
 
 ## [1.14.1] - 2020-12-15